fixed a lot of stuff. a whole lot.
Now detects if there is an additional GPU and displays it properly if so. Also removed some garbage like (R) and (TM) from the script's output. This saves space in the terminal and looks better. Also made clear in the script's comments that this is GPL'd software now, not MIT licensed.